Hack for Table of Contents ToC positioning
This forces the ToC into the .prose div as a child of the post content.
Doing this allows images and other content to break out of the grid column.
This hack moves the ToC into the right column, bumps it to the top (grid-row: 1/ -1),
and shrinks it's height so it whatever content element in the same row doesn't take
up any space.